requirements.txt
absl-py==1.4.0
aiofiles==22.1.0
aiohttp==3.8.4
aiosignal==1.3.1
aiosqlite==0.18.0
albumentations==1.3.1
annotated-types==0.5.0
antlr4-python3-runtime==4.9.3
anyio==3.5.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.0.5
async-lru==2.0.2
async-timeout==4.0.2
attrs==22.1.0
av==10.0.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==4.1.0
cachetools==5.3.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.7
cmake==3.25.0
comm==0.1.2
contourpy==1.0.7
cycler==0.11.0
Cython==0.29.36
datasets==2.13.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
diffusers==0.18.2
dill==0.3.6
docopt==0.6.2
dominate==2.8.0
einops==0.6.1
entrypoints==0.4
executing==0.8.3
fastapi==0.100.0
fastjsonschema==2.16.2
filelock==3.9.0
flit_core==3.6.0
fonttools==4.39.2
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.6.0
gensim==4.3.2
google-auth==2.22.0
google-auth-oauthlib==1.0.0
graphviz==0.20.1
grpcio==1.56.0
hdbscan==0.8.33
huggingface-hub==0.15.1
idna==3.4
imageio==2.31.4
importlib-metadata==6.8.0
invisible-watermark==0.2.0
ipykernel==6.19.2
ipython==8.10.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.11
jsonpatch==1.33
jsonpointer==2.3
jsonschema==4.17.3
jupyter_client==7.4.9
jupyter_core==5.2.0
jupyter-events==0.6.3
jupyter-lsp==2.2.0
jupyter_server==2.5.0
jupyter_server_fileid==0.8.0
jupyter_server_terminals==0.4.4
jupyter_server_ydoc==0.8.0
jupyter-ydoc==0.2.3
jupyterlab==4.0.6
jupyterlab-pygments==0.1.2
jupyterlab_server==2.20.0
jupyterlab-widgets==3.0.7
kiwisolver==1.4.4
latexify-py==0.2.0
lazy_loader==0.3
ldm-fix==1.0.1
lightning-utilities==0.9.0
lit==15.0.7
llvmlite==0.41.0
lxml==4.9.1
Markdown==3.4.3
MarkupSafe==2.1.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==2.0.5
mpmath==1.2.1
multidict==6.0.4
multiprocess==0.70.14
nb-conda-kernels==2.3.1
nbclassic==0.5.2
nbclient==0.5.13
nbconvert==7.2.10
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==3.0
nltk==3.8.1
notebook==6.5.2
notebook_shim==0.2.2
numba==0.58.0
numpy==1.24.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.7.0.72
opencv-python-headless==4.8.1.78
overrides==7.3.1
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pip==23.2.1
pipreqs==0.4.13
platformdirs==2.5.2
prometheus-client==0.14.1
prompt-toolkit==3.0.36
protobuf==4.23.4
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.0.3
pydantic_core==2.3.0
Pygments==2.11.2
pynndescent==0.5.10
pyparsing==3.0.9
pyrsistent==0.18.0
python-dateutil==2.8.2
python-json-logger==2.0.7
pytorch-lightning==2.0.5
pytz==2022.7.1
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.0.1
qudida==0.0.4
regex==2023.6.3
requests==2.28.1
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
safetensors==0.3.1
scikit-image==0.21.0
scikit-learn==1.3.1
scipy==1.11.1
seaborn==0.12.2
Send2Trash==1.8.0
setuptools==65.6.3
six==1.16.0
smart-open==6.4.0
sniffio==1.2.0
soupsieve==2.3.2.post1
stack-data==0.2.0
starlette==0.27.0
sympy==1.11.1
tbb==2021.10.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
terminado==0.17.1
textblob==0.17.1
threadpoolctl==3.2.0
tifffile==2023.9.26
tinycss2==1.2.1
tokenizers==0.13.3
tomli==2.0.1
top2vec==1.0.29
torch==2.0.0
torchaudio==2.0.1
torchmetrics==1.0.0
torchvision==0.15.1
torchviz==0.0.2
tornado==6.2
tqdm==4.65.0
traitlets==5.7.1
transformers==4.31.0
triton==2.0.0
typing_extensions==4.7.1
umap-learn==0.5.4
uri-template==1.2.0
urllib3==1.26.13
visdom==0.2.4
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==0.58.0
Werkzeug==2.3.6
wheel==0.38.4
widgetsnbextension==4.0.7
wikipedia==1.4.0
wordcloud==1.9.2
xxhash==3.2.0
y-py==0.5.9
yarg==0.1.9
yarl==1.9.2
ypy-websocket==0.8.2
zipp==3.16.2
